requirements.txt
python-docx
PyPDF2
requests
